pyproject.toml
[project]
name = "guided-modality-adapter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.12.0",
    "omegaconf>=2.3.0",
    "openai-whisper>=20250625",
    "peft>=0.18.0",
    "pyannote-audio>=4.0.2",
    "streamlit>=1.52.1",
    "torch>=2.8.0",
    "transformers>=4.57.3",
]
